refactor: 회원가입 폼 리팩토링
